Split FavoritesScreen into a favorites package
FavoritesScreen.kt was 1755 lines holding six unrelated concerns: the
row/state models, the ViewModel, the screen and its content, the sense
card, the two study cards, and eighteen previews with their sample
builders. ui/word/ and ui/study/ were already laid out the way this
should be; this applies the same shape to Favorites.

  ui/favorites/FavoritesViewModel.kt      models, UiState, ViewModel
  ui/favorites/FavoritesScreen.kt         Screen, Content, empty states
  ui/favorites/FavoriteSenseCard.kt       the sense row
  ui/favorites/StudyCards.kt              StudyDueCard, StudyDoneCard
  ui/favorites/preview/FavoritesSamples.kt   createMockSense/createSenseItem
  ui/favorites/preview/FavoritesPreviews.kt  the eighteen previews

Nothing else moves: ui/ keeps AppNavigationBar, Preview, SensePrefetch,
VoiceSetupBottomSheet and DeveloperFormatters, which Favorites now
imports explicitly rather than picking up from the same package.

This is a pure move. Diffing the moved bodies against the original file
shows them byte-identical apart from five deliberate widenings, each
required because the symbol now crosses a file boundary:

  FavoriteSenseCard, StudyDoneCard, StudyDueCard  private -> internal
  createMockSense, createSenseItem                private -> internal

Callers pick up the new package: App.kt adds a wildcard import beside its
existing ui one, ListDetailScreen imports toFavoriteSenseLoadError, and
FavoritesViewModelTest imports the package it exercises.
